Finish coding Stage
Got slot parenting to work, and now prefabs including a couple components now
Import

The huge issue I have is getting the slots to merge with FBX imports. May need to implement a copy of the froox engine fbx importer just so I can hook into the right setup slot events.
Because idk how to merge the generated prefab with the FBX that imports much much later and is an ASYNC